+++ frontend/src/css/TodayCom.css
@@ -1,0 +1,32 @@
+.todolist {
+  width: 618px;
+  height: 750px;
+}
+.todolist_index {
+  height: 70px;
+  font-size: 36px;
+  padding-left: 30px;
+  margin-top: 10px;
+  font-weight: 550;
+}
+.todolist_content {
+  width: 618px;
+  height: 640px;
+  background-color: #fefcfc;
+  border-radius: 30px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+.todolist_checkbox {
+  height: 350px;
+  width: 500px;
+}
+.todolist_memo {
+  height: 200px;
+  width: 580px;
+  border-radius: 30px;
+  background-color: #efefef;
+  padding: 20px;
+}
